Backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn",
    "passlib[bcrypt]",
    "pyjwt",
    "python-dotenv",
    "sqlalchemy>=2.0.39",
    "sqlalchemy-utils>=0.41.2",
    "pydantic-settings>=2.8.1",
    "pymysql>=1.1.1",
    "pydantic[email]>=2.10.6",
    "bcrypt==4.0.1",
    "requests>=2.32.3",
    "flake8>=7.1.2",
    "pylint>=3.3.6",
    "black>=25.1.0",
    "isort>=6.0.1",
    "ruff>=0.11.2",
    "cloudinary>=1.43.0",
    "python-multipart>=0.0.20",
]

